backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
python-docx>=1.1.0
psutil>=6.1.0
sse-starlette>=2.2.0
python-multipart>=0.0.20
bcrypt>=4.2.0
PyJWT>=2.9.0
